+++ tockn/view/src/components/GistCard.vue
@@ -1,25 +1,48 @@
 <template>
- <div class="card">
-   <div class="title">
-     {{ title }}
-   </div>
-   <div class="content">
-     {{ content }}
-   </div>
- </div>
+  <router-link class="a" v-bind:to="`/gists/${gist.id}`" >
+    <div class="card">
+      <div class="title">
+        <p>{{ cutDesc(gist.description) }}</p>
+      </div>
+      <div class="content">
+        <p v-for="(name, index) in fileNames(gist.files)" :key="index">{{ name }}</p>
+      </div>
+    </div>
+  </router-link>
 </template>
 
 <script>
 
 export default {
   props: {
-    title: String,
-    content: String
+    gist: Object
+  },
+  methods: {
+    cutDesc (text) {
+      if (text === '' || text === null) {
+        return '-No Description-'
+      }
+      if (text.length > 100) {
+        return text.slice(0, 100) + '...'
+      }
+      return text
+    },
+    fileNames (files) {
+      let names = Object.keys(files)
+      if (names.length > 10) {
+        names = names.slice(0, 10)
+        names.push('...and more!')
+      }
+      return names
+    }
   }
 }
 
 </script>
 
-<style>
-
+<style scoped>
+.a {
+  text-decoration: none;
+  color: black;
+}
 </style>
